+++ lib/styles.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+const textStyle = TextStyle(
+  fontFamily: 'Roboto',
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: Colors.green,
+);
+
+const textStyle1 = TextStyle(
+  fontFamily: 'Roboto',
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: const Color(0xFF3B3E5B),
+);
+
+const textStyle2 = TextStyle(
+  fontFamily: 'Roboto',
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: Colors.yellow,
+);
+
+const textStyle3 = TextStyle(
+  fontFamily: 'Roboto',
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: const Color(0xFF3B3E5B),
+);
